lambda/requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
attrs==19.1.0
autopep8==1.4.4
aws==0.2.5
awscli==1.16.193
bcrypt==3.1.7
beautifulsoup4==4.7.1
black==19.3b0
boto==2.49.0
botocore==1.12.183
cffi==1.12.3
Click==7.0
colorama==0.3.9
cryptography==2.7
docutils==0.14
entrypoints==0.3
fabric==2.4.0
flake8==3.7.7
flake8-import-order==0.18.1
invoke==1.2.0
isort==4.3.21
jmespath==0.9.4
mccabe==0.6.1
paramiko==2.6.0
prettytable==0.7.2
pyasn1==0.4.5
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyNaCl==1.3.0
pynamodb==3.4.1
python-dateutil==2.8.0
python-dotenv==0.10.3
PyYAML==5.1
rsa==3.4.2
s3transfer==0.2.1
six==1.12.0
soupsieve==1.9.2
toml==0.10.0
urllib3==1.25.3
